+++ 14.24.3_0/js/help.js
@@ -1,1 +1,0 @@
-var c=new URL(location.href).searchParams.get("src")||"";docs.src=c||"https://docs.lightningautofill.com/";